src_py/pyproject.toml
[project]
name = "agent-adapter"
version = "0.1.0"
description = "Agent Adapter package"
requires-python = ">=3.11"
dependencies = ["google-genai>=1.5.0", "httpx[socks]"]

[dependency-groups]
dev = ["pytest>=8.4.2", "pytest-asyncio>=0.23.0", "ruff>=0.14.3"]

[build-system]
requires = ["uv_build>=0.9.16,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""
source-include = []

[tool.ruff]
target-version = "py311"
line-length = 119
indent-width = 4

[tool.ruff.lint]
ignore = ["C901", "E501", "E741", "W605", "C408"]
select = ["C", "E", "F", "I", "W", "RUF022"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403", "F811"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
